+++ code/main.py
@@ -1,9 +1,8 @@
-from langchain.document_loaders import PyPDFLoader, DirectoryLoader
+from langchain_community.document_loaders import PyPDFLoader, DirectoryLoader
 from langchain import PromptTemplate
-from langchain.embeddings import HuggingFaceEmbeddings
-from langchain.vectorstores import FAISS
+from langchain_community.embeddings import HuggingFaceEmbeddings
+from langchain_community.vectorstores import FAISS
 from langchain.chains import RetrievalQA
-from langchain.llms import CTransformers
 import chainlit as cl
 from langchain_community.chat_models import ChatOpenAI
 from langchain_community.embeddings import OpenAIEmbeddings
@@ -11,13 +10,22 @@ import yaml
 import logging
 from dotenv import load_dotenv
 
-from modules.llm_tutor import LLMTutor
-from modules.constants import *
-from modules.helpers import get_sources
+import os
+import sys
+
+# Add the 'code' directory to the Python path
+current_dir = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(current_dir)
+
+from modules.chat.llm_tutor import LLMTutor
+from modules.config.constants import *
+from modules.chat.helpers import get_sources
 
 
+global logger
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
+logger.propagate = False
 
 # Console Handler
 console_handler = logging.StreamHandler()
@@ -25,13 +33,6 @@ console_handler.setLevel(logging.INFO)
 formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
 console_handler.setFormatter(formatter)
 logger.addHandler(console_handler)
-
-# File Handler
-log_file_path = "log_file.log"  # Change this to your desired log file path
-file_handler = logging.FileHandler(log_file_path)
-file_handler.setLevel(logging.INFO)
-file_handler.setFormatter(formatter)
-logger.addHandler(file_handler)
 
 
 # Adding option to select the chat profile
@@ -66,12 +67,26 @@ def rename(orig_author: str):
 # chainlit code
 @cl.on_chat_start
 async def start():
-    with open("code/config.yml", "r") as f:
+    with open("modules/config/config.yml", "r") as f:
         config = yaml.safe_load(f)
-        print(config)
-        logger.info("Config file loaded")
-        logger.info(f"Config: {config}")
-        logger.info("Creating llm_tutor instance")
+
+    # Ensure log directory exists
+    log_directory = config["log_dir"]
+    if not os.path.exists(log_directory):
+        os.makedirs(log_directory)
+
+    # File Handler
+    log_file_path = (
+        f"{log_directory}/tutor.log"  # Change this to your desired log file path
+    )
+    file_handler = logging.FileHandler(log_file_path, mode="w")
+    file_handler.setLevel(logging.INFO)
+    file_handler.setFormatter(formatter)
+    logger.addHandler(file_handler)
+
+    logger.info("Config file loaded")
+    logger.info(f"Config: {config}")
+    logger.info("Creating llm_tutor instance")
 
     chat_profile = cl.user_session.get("chat_profile")
     if chat_profile is not None:
@@ -93,8 +108,7 @@ async def start():
     llm_tutor = LLMTutor(config, logger=logger)
 
     chain = llm_tutor.qa_bot()
-    model = config["llm_params"]["local_llm_params"]["model"]
-    msg = cl.Message(content=f"Starting the bot {model}...")
+    msg = cl.Message(content=f"Starting the bot {chat_profile}...")
     await msg.send()
     msg.content = opening_message
     await msg.update()
@@ -104,24 +118,17 @@ async def start():
 
 @cl.on_message
 async def main(message):
+    global logger
     user = cl.user_session.get("user")
     chain = cl.user_session.get("chain")
-    # cb = cl.AsyncLangchainCallbackHandler(
-    #     stream_final_answer=True, answer_prefix_tokens=["FINAL", "ANSWER"]
-    # )
-    # cb.answer_reached = True
-    # res=await chain.acall(message, callbacks=[cb])
-    res = await chain.acall(message.content)
-    print(f"response: {res}")
+    cb = cl.AsyncLangchainCallbackHandler()  # TODO: fix streaming here
+    cb.answer_reached = True
+    res = await chain.acall(message.content, callbacks=[cb])
+    # res = await chain.acall(message.content)
     try:
         answer = res["answer"]
     except:
         answer = res["result"]
-    print(f"answer: {answer}")
-
-    logger.info(f"Question: {res['question']}")
-    logger.info(f"History: {res['chat_history']}")
-    logger.info(f"Answer: {answer}\n")
 
     answer_with_sources, source_elements = get_sources(res, answer)
