+++ code/modules/config/config.yml
@@ -1,23 +1,32 @@
-embedding_options:
+log_dir: '../storage/logs' # str
+log_chunk_dir: '../storage/logs/chunks' # str
+device: 'cpu' # str [cuda, cpu]
+
+vectorstore:
   embedd_files: False # bool
-  data_path: 'storage/data' # str
-  url_file_path: 'storage/data/urls.txt' # str
-  expand_urls: True # bool
-  db_option : 'RAGatouille' # str [FAISS, Chroma, RAGatouille]
-  db_path : 'vectorstores' # str
+  data_path: '../storage/data' # str
+  url_file_path: '../storage/data/urls.txt' # str
+  expand_urls: False # bool
+  db_option : 'Chroma' # str [FAISS, Chroma, RAGatouille]
+  db_path : '../vectorstores' # str
   model : 'sentence-transformers/all-MiniLM-L6-v2' # str [sentence-transformers/all-MiniLM-L6-v2, text-embedding-ada-002']
   search_top_k : 3 # int
   score_threshold : 0.2 # float
+
+  faiss_params: # Not used as of now
+    index_path: '../vectorstores/faiss.index' # str
+    index_type: 'Flat' # str [Flat, HNSW, IVF]
+    index_dimension: 384 # int
+    index_nlist: 100 # int
+    index_nprobe: 10 # int
+
 llm_params: 
   use_history: True # bool
   memory_window: 3 # int
   llm_loader: 'openai' # str [local_llm, openai]
   openai_params:
     model: 'gpt-3.5-turbo-1106' # str [gpt-3.5-turbo-1106, gpt-4]
-  local_llm_params:
-    model: "storage/models/llama-2-7b-chat.Q4_0.gguf"
-    model_type: "llama"
-    temperature: 0.2
+
 splitter_options:
   use_splitter: True # bool
   split_by_token : True # bool
