+++ code/modules/retriever/__init__.py
@@ -1,0 +1,2 @@
+from .faiss_retriever import FaissRetriever
+from .chroma_retriever import ChromaRetriever
