+++ code/modules/retriever/base.py
@@ -1,0 +1,6 @@
+class BaseRetriever:
+    def __init__(self, config):
+        self.config = config
+
+    def return_retriever(self):
+        raise NotImplementedError
