+++ code/modules/vectorstore/__init__.py
@@ -1,0 +1,2 @@
+from .base import VectorStoreBase
+from .faiss import FAISS
