+++ code/modules/vectorstore/embedding_model_loader.py
@@ -2,10 +2,7 @@ from langchain_community.embeddings import OpenAIEmbeddings
 from langchain_community.embeddings import HuggingFaceEmbeddings
 from langchain_community.embeddings import LlamaCppEmbeddings
 
-try:
-    from modules.constants import *
-except:
-    from constants import *
+from modules.config.constants import *
 import os
 
 
@@ -14,19 +11,19 @@ class EmbeddingModelLoader:
         self.config = config
 
     def load_embedding_model(self):
-        if self.config["embedding_options"]["model"] in ["text-embedding-ada-002"]:
+        if self.config["vectorstore"]["model"] in ["text-embedding-ada-002"]:
             embedding_model = OpenAIEmbeddings(
                 deployment="SL-document_embedder",
-                model=self.config["embedding_options"]["model"],
+                model=self.config["vectorestore"]["model"],
                 show_progress_bar=True,
                 openai_api_key=OPENAI_API_KEY,
                 disallowed_special=(),
             )
         else:
             embedding_model = HuggingFaceEmbeddings(
-                model_name=self.config["embedding_options"]["model"],
+                model_name=self.config["vectorstore"]["model"],
                 model_kwargs={
-                    "device": "cpu",
+                    "device": f"{self.config['device']}",
                     "token": f"{HUGGINGFACE_TOKEN}",
                     "trust_remote_code": True,
                 },
